Windows/MSVC portability fixes
Build and run MB-System under MSVC without affecting Linux/Mac builds.
All Windows-specific code is guarded by _WIN32 / _MSC_VER / MSVC.

- mb_define.h: minimal XDR shim when no SunRPC is available on Windows
- mb_io.h: __declspec(dllexport/dllimport) markers for sensor-type data
  symbols (CMAKE_WINDOWS_EXPORT_ALL_SYMBOLS only covers functions)
- mb_mem.c: intptr_t instead of long for pointer values (64-bit safe);
  %zx for size_t formatting
- mbsys_kmbes.c: drop local Win32 clock_gettime() (provided globally via
  mb_win_compat.h); removed block also had a non-compiling "return = -1"
- CMakeLists: link MB_SYSLIBS instead of hardcoded m/pthread; MSVC /STACK
  bump (8 MB) to avoid stack overflow on piped stdin; force-include
  mb_win_compat.h on MSVC
- include/mb_win_compat.h, include/unistd.h: Windows shim headers
- misc portability cleanups across gsf, surf, bsio, mbtrn*, utilities

+++ src/mbio/mbr_gsfgenmb.c
@@ -381,7 +381,7 @@ fprintf(stderr, "dataID->recordID: %d  Read GSF_RECORD_ATTITUDE\n", dataID->reco
 		fprintf(stderr, "\ndbg4  New comment read by MBIO function <%s>\n", __func__);
 		fprintf(stderr, "dbg4  New comment values:\n");
 		fprintf(stderr, "dbg4       kind:              %d\n", data->kind);
-		fprintf(stderr, "dbg4       comment time sec:  %ld\n", records->comment.comment_time.tv_sec);
+		fprintf(stderr, "dbg4       comment time sec:  %lld\n", (long long)records->comment.comment_time.tv_sec);
 		fprintf(stderr, "dbg4       comment time nsec: %ld\n", records->comment.comment_time.tv_nsec);
 		fprintf(stderr, "dbg4       comment length:    %d\n", records->comment.comment_length);
 		fprintf(stderr, "dbg4       comment:           %s\n", records->comment.comment);
@@ -390,7 +390,7 @@ fprintf(stderr, "dataID->recordID: %d  Read GSF_RECORD_ATTITUDE\n", dataID->reco
 		fprintf(stderr, "\ndbg4  New ping read by MBIO function <%s>\n", __func__);
 		fprintf(stderr, "dbg4  New ping values:\n");
 		fprintf(stderr, "dbg4       kind:       %d\n", data->kind);
-		fprintf(stderr, "dbg4       time sec:   %ld\n", mb_ping->ping_time.tv_sec);
+		fprintf(stderr, "dbg4       time sec:   %lld\n", (long long)mb_ping->ping_time.tv_sec);
 		fprintf(stderr, "dbg4       time nsec:  %ld\n", mb_ping->ping_time.tv_nsec);
 		fprintf(stderr, "dbg4       longitude:  %f\n", mb_ping->longitude);
 		fprintf(stderr, "dbg4       latitude:   %f\n", mb_ping->latitude);
@@ -468,7 +468,7 @@ int mbr_wt_gsfgenmb(int verbose, void *mbio_ptr, void *store_ptr, int *error) {
 		fprintf(stderr, "\ndbg4  New comment to be written by MBIO function <%s>\n", __func__);
 		fprintf(stderr, "dbg4  New comment values:\n");
 		fprintf(stderr, "dbg4       kind:              %d\n", data->kind);
-		fprintf(stderr, "dbg4       comment time sec:  %ld\n", records->comment.comment_time.tv_sec);
+		fprintf(stderr, "dbg4       comment time sec:  %lld\n", (long long)records->comment.comment_time.tv_sec);
 		fprintf(stderr, "dbg4       comment time nsec: %ld\n", records->comment.comment_time.tv_nsec);
 		fprintf(stderr, "dbg4       comment length:    %d\n", records->comment.comment_length);
 		fprintf(stderr, "dbg4       comment:           %s\n", records->comment.comment);
@@ -477,7 +477,7 @@ int mbr_wt_gsfgenmb(int verbose, void *mbio_ptr, void *store_ptr, int *error) {
 		fprintf(stderr, "\ndbg4  New ping to be written by MBIO function <%s>\n", __func__);
 		fprintf(stderr, "dbg4  New ping values:\n");
 		fprintf(stderr, "dbg4       kind:       %d\n", data->kind);
-		fprintf(stderr, "dbg4       time sec:   %ld\n", mb_ping->ping_time.tv_sec);
+		fprintf(stderr, "dbg4       time sec:   %lld\n", (long long)mb_ping->ping_time.tv_sec);
 		fprintf(stderr, "dbg4       time nsec:  %ld\n", mb_ping->ping_time.tv_nsec);
 		fprintf(stderr, "dbg4       longitude:  %f\n", mb_ping->longitude);
 		fprintf(stderr, "dbg4       latitude:   %f\n", mb_ping->latitude);
