+++ src/utilities/mbnavlist.cc
@@ -811,7 +811,7 @@ int main(int argc, char **argv) {
 							case 'U': /* unix time in seconds since 1/1/70 00:00:00 */
 								time_u = (int)time_d;
 								if (ascii) {
-									printf("%ld", time_u);
+									printf("%lld", (long long)time_u);
 								} else {
 									double b = time_u;
 									fwrite(&b, sizeof(double), 1, stdout);
@@ -824,7 +824,7 @@ int main(int argc, char **argv) {
 									first_u = false;
 								}
 								if (ascii) {
-									printf("%ld", time_u - time_u_ref);
+									printf("%lld", (long long)(time_u - time_u_ref));
 								} else {
 									double b = time_u - time_u_ref;
 									fwrite(&b, sizeof(double), 1, stdout);
