+++ src/utilities/mbsegylist.cc
@@ -623,7 +623,7 @@ int main(int argc, char **argv) {
 				case 'U': /* unix time in seconds since 1/1/70 00:00:00 */
 					time_u = (int)time_d;
 					if (ascii)
-						printf("%ld", time_u);
+						printf("%lld", (long long)time_u);
 					else {
 						const double b = time_u;
 						fwrite(&b, sizeof(double), 1, stdout);
@@ -636,7 +636,7 @@ int main(int argc, char **argv) {
 						first_u = false;
 					}
 					if (ascii)
-						printf("%ld", time_u - time_u_ref);
+						printf("%lld", (long long)(time_u - time_u_ref));
 					else {
 						const double b = time_u - time_u_ref;
 						fwrite(&b, sizeof(double), 1, stdout);
